backend/Hygieiora_Backend/requirements.txt
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.0.0
constantly==23.10.4
cryptography==42.0.1
daphne==4.0.0
Django==4.2.9
django-cors-headers==4.3.1
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.0.0
python-decouple==3.8
pytz==2023.3.post1
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing-extensions==4.9.0
tzdata==2023.4
zope.interface==6.1
